requirements.txt
requests
faker
pytest
kafka-python
psycopg2-binary